paint-gradient-editor: extract StopColorPicker to its own module; add light/dark mode CSS
Move the stop colour picker into stop-color-picker.js as a standalone ES
module with all dependencies injected via the constructor.  No imports —
the module is pure logic with no coupling to the outer userscript.

Replace inline <style> injection with static CSS classes in style.css.
Picker panel colours now use --editorDarkMode-* variables with Scratch
light-mode fallbacks, giving automatic light/dark mode support that matches
the native colour picker popover.
