+++ addons/paint-gradient-editor/style.css
@@ -84,3 +84,146 @@
   font-size: 0.75rem;
   color: #575e75;
 }
+
+/* ── Extra gradient stop colour picker popup ────────────────────────────── */
+/* Colours use editor-dark-mode variables with light-mode Scratch fallbacks, */
+/* mirroring the same pattern used by the native .Popover-body colour picker. */
+
+.sa-extra-stop-picker {
+  position: fixed;
+  border-radius: 7px;
+  padding: 10px;
+  z-index: 99999;
+  user-select: none;
+  width: 200px;
+  box-sizing: content-box;
+  overflow: hidden;
+  background-color: var(--editorDarkMode-input, white);
+  border: 1px solid var(--editorDarkMode-border, rgba(0, 0, 0, 0.15));
+  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.22);
+  color: var(--editorDarkMode-input-text, #575e75);
+}
+
+.sa-stop-picker-handle {
+  margin: -10px -10px 10px;
+  padding: 5px 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: grab;
+  flex-shrink: 0;
+  background-color: rgba(0, 0, 0, 0.08);
+}
+
+.sa-stop-picker-handle:active {
+  cursor: grabbing;
+}
+
+.sa-stop-picker-grip {
+  display: grid;
+  grid-template-columns: repeat(3, 4px);
+  gap: 3px;
+  opacity: 0.45;
+}
+
+.sa-stop-picker-dot {
+  width: 3px;
+  height: 3px;
+  border-radius: 50%;
+  background-color: var(--editorDarkMode-input-text, #575e75);
+}
+
+.sa-stop-picker-sv {
+  border-radius: 4px;
+  cursor: crosshair;
+  display: block;
+}
+
+.sa-stop-picker-thumb {
+  position: absolute;
+  top: 0;
+  width: 18px;
+  height: 18px;
+  border-radius: 50%;
+  background: white;
+  box-shadow:
+    0 0 0 2px rgba(0, 0, 0, 0.45),
+    0 1px 4px rgba(0, 0, 0, 0.6);
+  transform: translateX(-50%);
+  pointer-events: none;
+}
+
+.sa-stop-picker-alpha-fill {
+  position: absolute;
+  inset: 0;
+  border-radius: 6px;
+  pointer-events: none;
+}
+
+.sa-stop-picker-bottom-row {
+  display: flex;
+  align-items: center;
+  gap: 6px;
+  margin-top: 8px;
+}
+
+.sa-stop-picker-swatch {
+  width: 22px;
+  height: 22px;
+  border-radius: 3px;
+  flex-shrink: 0;
+  border: 1px solid var(--editorDarkMode-border, rgba(0, 0, 0, 0.15));
+}
+
+.sa-stop-picker-swatch-fill {
+  width: 100%;
+  height: 100%;
+  border-radius: 2px;
+}
+
+.sa-stop-picker-input {
+  background-color: var(--editorDarkMode-input, white);
+  color: var(--editorDarkMode-input-text, #575e75);
+  border: 1px solid var(--editorDarkMode-border, rgba(0, 0, 0, 0.15));
+  border-radius: 3px;
+  outline: none;
+  box-sizing: border-box;
+}
+
+.sa-stop-picker-hex {
+  flex: 1;
+  min-width: 0;
+  padding: 3px 5px;
+  font-family: monospace;
+  font-size: 12px;
+}
+
+.sa-stop-picker-alpha-input {
+  width: 42px;
+  padding: 3px 4px;
+  font-size: 11px;
+  text-align: right;
+  flex-shrink: 0;
+}
+
+.sa-stop-picker-alpha-label {
+  font-size: 11px;
+  font-family: sans-serif;
+  flex-shrink: 0;
+  opacity: 0.6;
+}
+
+.sa-stop-picker-eyedropper {
+  width: 22px;
+  height: 22px;
+  padding: 0;
+  border-radius: 3px;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-shrink: 0;
+  background-color: var(--editorDarkMode-input, white);
+  border: 1px solid var(--editorDarkMode-border, rgba(0, 0, 0, 0.15));
+  color: var(--editorDarkMode-input-text, #575e75);
+}
